Set OnAttackSameSpecie delegate in Enemy class
